+++ app/Http/Controllers/Admin/StateController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\State;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class StateController extends Controller
 {
@@ -23,7 +24,7 @@ class StateController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -32,7 +33,12 @@ class StateController extends Controller
             'name' => 'required'
         ]);
 
-        State::create($request->only('name'));
+        if($request->get('is_resolved')){
+            State::create(['name' => $request->name, 'is_resolved' => 1]);
+        }
+        else {
+            State::create($request->only('name'));
+        }
 
         return redirect()->back();
     }
@@ -40,7 +46,7 @@ class StateController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\State  $state
+     * @param \App\Models\State $state
      * @return \Illuminate\Http\Response
      */
     public function edit(State $state)
@@ -51,8 +57,8 @@ class StateController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\State  $state
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\State $state
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, State $state)
@@ -61,7 +67,12 @@ class StateController extends Controller
             'name' => 'required'
         ]);
 
-        $state->update($request->only('name'));
+        if($request->get('is_resolved')) {
+            $state->update(['name' => $request->name, 'is_resolved' => 1]);
+        }
+        else {
+            $state->update($request->only('name'));
+        }
 
         return redirect()->back();
     }
@@ -69,11 +80,15 @@ class StateController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\State  $state
+     * @param \App\Models\State $state
      * @return \Illuminate\Http\Response
      */
     public function destroy(State $state)
     {
+        if (State::count() <= 1){
+            throw new BadRequestHttpException("The last one state.");
+        }
+
         $state->delete();
 
         return redirect()->back();
