+++ app/Http/Controllers/Admin/TypeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Type;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class TypeController extends Controller
 {
@@ -74,6 +75,10 @@ class TypeController extends Controller
      */
     public function destroy(Type $type)
     {
+        if (Type::count() <= 1){
+            throw new BadRequestHttpException("The last one type.");
+        }
+
         $type->delete();
 
         return redirect()->back();
